refactor UI out
